Modify collection ingest/update to track the master branch
Collections are now detected by the presence of
`"keywords": ["element-collection"]` in `master/bower.json`.

Library entities for collections maintain a sequence number
that is incremented each time we detect that the master branch
has changed. The sequence number is used to construct
artificial version tags which can be compared appropriately.