Suppress bower equivalent when publishing NPM package (#1118)
When publishing an NPM package, we may find an equivalent repo that was published under Bower. As we create a new `Library` instance to represent the NPM package, we need to remove the Bower repo to remove duplication.

Deleting a repo in this case represents an attack vector, as we use the GitHub repo specified in `package.json`, which may be malicious. To achieve the same behavior without permanent damage, suppression is used. Entities under suppression will not appear across the site, but remain intact and can be easily restored.

The reason it is a separate task is because the existing architecture requires singular database transactions. Since this acts on a different `Library` entity, it needs to be pulled into its own transactional task.