Filter analysis result (#1124)
Filtering files for analysis was not sufficient. If a file depends on a file that was not found in the search, it can still end up in the analyzed output. This change filters results before final output of analyzing results. Filtering before is still useful as that forms analyzer's search roots and can help prevent some additional overhead associated with having to analyze additional files.