Add workflow that automatically adds new webcomponents.org issues to the lit project board (#1426)
I'm not 100% sure if this will work yet. It is the same as the one we use at https://github.com/lit/lit/blob/main/.github/workflows/add-issues-to-project.yml. The last time I tried this, the API call here failed because the issue was not in the same org as the project. However, according to https://github.com/orgs/community/discussions/6212#discussioncomment-1781562 this may now work.  I had to create a new lit-robot automation PAT because the old one was not saved anywhere. The new one is saved in the appropriate internal place. Let's see what happens!

If this works, as a followup I will have it automatically add the webcomponents.org category